// frontend/src/api/models/listTenantsParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { ListTenantsPlan } from './listTenantsPlan'
import type { ListTenantsSortBy } from './listTenantsSortBy'
import type { ListTenantsSortDir } from './listTenantsSortDir'
import type { ListTenantsStatus } from './listTenantsStatus'

export type ListTenantsParams = {
  /**
   * Search keyword
   */
  keyword?: string
  /**
   * Tenant status
   */
  status?: ListTenantsStatus
  /**
   * Subscription plan
   */
  plan?: ListTenantsPlan
  /**
   * Page number
   */
  page?: number
  /**
   * Items per page
   * @maximum 100
   */
  page_size?: number
  /**
   * Sort by field
   */
  sort_by?: ListTenantsSortBy
  /**
   * Sort direction
   */
  sort_dir?: ListTenantsSortDir
}