// frontend/src/api/models/listTenantsStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListTenantsStatus = (typeof ListTenantsStatus)[keyof typeof ListTenantsStatus]

export const ListTenantsStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
  trial: 'trial',
} as const