// frontend/src/api/models/listTenantsSortBy.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListTenantsSortBy = (typeof ListTenantsSortBy)[keyof typeof ListTenantsSortBy]

export const ListTenantsSortBy = {
  code: 'code',
  name: 'name',
  status: 'status',
  plan: 'plan',
  created_at: 'created_at',
  updated_at: 'updated_at',
} as const